Version change to 0.99.4
Updated all calls to read_csv() to read in as character.

Adjusted the function grading subroutine to account for everything coming in as character.

Choose to suppress warnings in the regex grading in the event there are no matches for a particular regex, perhaps this should be rolled back? Will the reader need this as an early indication there are no matches? Could be confusing to keep it as an indicator if there is a partial credit regex match which no one gets.

Updated the class of "possible" to be character in make_report.